manageOrder.html and styleTable.css added
Manage product can see all the order list. The admin can accept or reject the order. If he accept the order he will get the information about the customers. But if he reject the order, he can send a message to the customer to explain why he rejected the order.